Create makePlayer() and makeDargon() functions
having a function that returns a subtype of a creature as
a Creature object does not work as well as I initialy thought.

If the player is a Creature, then the use of functions in
Player.cpp are unable to be used, somewhat destroying the
convenience of the factory.

this will probably be a temporary fix and changed in a later
commit