Define Coordinate struct and Coordinate compass
Coordinates are defined as an x and y position on the map.
Also defined in the struct are overloaded versions of
many default operators, such as +, -, ==, and so on.

The compass is used in conjuction with the directions enum
defined in Enums.h. This allows for easy and simple modification
of an actor or items Coordinates.