Change move to moveOrAttack
move now also checks if the player should attack a creatue. The NEW_TURN
state of the game is also changed in here.

Before when the game state was changing if the player moved, the
monsters would not be updated. This was because NEW_TURN was being set
only when the player made a move, and not when they attacked.