Implement map search with Dijkstra
Can now compute a path between two coordinates with the use of a
Dijkstra. the Compute path simply gets the path from one coordinate to
the next. The walkPath function walks a node in the path only if it is
walkable. Right now a monster will not continue following the player if
they are out of its field of view. This is due to recomputing the path
everytime the monster tries to walk it.

Eventually, a custom callback will be used in order to make Actors
follow the path prefering to walk on the safest tiles.