Create class for the Gui
moved TCODConsole::root functions over to this new Gui class, as well as
adding a new status window that displays logs and player stats. also
displays stats of tile under cursor if it has a monster, item, or player

the logs work currently, but displaying player stats is not yet
implemented.